Add authorization

+++ vagrant/catalog/views.py
@@ -1,7 +1,8 @@
+#!/usr/bin/env python
 from flask import Flask, render_template, request, redirect, url_for
 from flask import flash, jsonify
-from sqlalchemy import create_engine
-from database_setup import Base, Category, Item
+from sqlalchemy import create_engine, asc
+from database_setup import Base, Category, Item, User
 from sqlalchemy.orm import sessionmaker
 
 # Imports for oauth2
@@ -16,7 +17,7 @@ from flask import make_response
 import requests
 
 engine = create_engine(
-    'sqlite:///catalog.db', connect_args={'check_same_thread': False})
+    'sqlite:///usersCatalog.db', connect_args={'check_same_thread': False})
 Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
@@ -44,7 +45,7 @@ def gconnect():
         response.headers['Content-Type'] = 'application/json'
         return response
     # Obtain authorization code
-    code = request.data #.decode('utf-8')
+    code = request.data  # .decode('utf-8')
 
     try:
         oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
@@ -113,11 +114,11 @@ def gconnect():
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
 
-    # # see if user exists, if it doesn't make a new one
-    # user_id = getUserID(login_session['email'])
-    # if not user_id:
-    #     user_id = createUser(login_session)
-    # login_session['user_id'] = user_id
+    # see if user exists, if it doesn't make a new one
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
 
     output = ''
     output += '<h1>Welcome, '
@@ -137,26 +138,62 @@ def gconnect():
 # User Helper Functions
 
 
-# def createUser(login_session):
-#     newUser = User(name=login_session['username'], email=login_session[
-#                    'email'], picture=login_session['picture'])
-#     session.add(newUser)
-#     session.commit()
-#     user = session.query(User).filter_by(email=login_session['email']).one()
-#     return user.id
-#
-#
-# def getUserInfo(user_id):
-#     user = session.query(User).filter_by(id=user_id).one()
-#     return user
-#
-#
-# def getUserID(email):
-#     try:
-#         user = session.query(User).filter_by(email=email).one()
-#         return user.id
-#     except:
-#         return None
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'], picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
+
+@app.route('/gdisconnect')
+def gdisconnect():
+    # Only disconnect a connected user.
+    access_token = login_session.get('access_token')
+    if access_token is None:
+        response = make_response(
+            json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Execute HTTP GET request to revoke current token
+    print 'user name: '
+    print login_session['username']
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print 'result: '
+    print result
+
+    if result['status'] == '200':
+        # Reset user session
+        del login_session['access_token']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+        response = make_response(
+            json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
 
 
 # Creating API Endpoints
@@ -182,14 +219,20 @@ def itemJSON(category_id, item_id):
 @app.route('/')
 @app.route('/categories/')
 def showAllCategories():
-    categories = session.query(Category).all()
-    return render_template('categories.html', categories=categories)
+    categories = session.query(Category).order_by(asc(Category.name))  # .all()
+    if 'username' not in login_session:
+        return render_template('publicCategories.html', categories=categories)
+    else:
+        return render_template('categories.html', categories=categories)
 
 
 @app.route('/categories/new/', methods=['GET', 'POST'])
 def createCategory():
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
-        newCategory = Category(name=request.form['name'])
+        newCategory = Category(
+            name=request.form['name'], user_id=login_session['user_id'])
         session.add(newCategory)
         session.commit()
         flash('New Category Created')
@@ -200,7 +243,15 @@ def createCategory():
 
 @app.route('/categories/<int:category_id>/edit/', methods=['GET', 'POST'])
 def editCategory(category_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     editedCategory = session.query(Category).filter_by(id=category_id).one()
+    if editedCategory.user_id != login_session['user_id']:
+        return "\
+        <script>\
+        function reject() {alert('Access denied.');}\
+        </script>\
+        <body onload='reject()'"
     if request.method == 'POST':
         if request.form['name']:
             editedCategory.name = request.form['name']
@@ -214,7 +265,15 @@ def editCategory(category_id):
 
 @app.route('/categories/<int:category_id>/delete/', methods=['GET', 'POST'])
 def deleteCategory(category_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     categoryToDelete = session.query(Category).filter_by(id=category_id).one()
+    if categoryToDelete.user_id != login_session['user_id']:
+        return "\
+        <script>\
+        function reject() {alert('Access denied.');}\
+        </script>\
+        <body onload='reject()'"
     if request.method == 'POST':
         session.delete(categoryToDelete)
         session.commit()
@@ -230,16 +289,32 @@ def deleteCategory(category_id):
 def showCategory(category_id):
     category = session.query(Category).filter_by(id=category_id).one()
     items = session.query(Item).filter_by(category_id=category.id)
-    return render_template('item.html', category=category, items=items)
+    creator = getUserInfo(category.user_id)
+    if 'username' not in login_session or creator.id != login_session['user_id']:
+        return render_template(
+            'publicItem.html', category=category, items=items, creator=creator)
+    else:
+        return render_template(
+            'item.html', category=category, items=items, creator=creator)
 
 
 @app.route('/categories/<int:category_id>/items/new/', methods=['GET', 'POST'])
 def createCategoryItem(category_id):
+    if 'username' not in login_session:
+        return redirect('/login')
+    category = session.query(Category).filter_by(id=category_id).one()
+    if category.user_id != login_session['user_id']:
+        return "\
+        <script>\
+        function reject() {alert('Access denied.');}\
+        </script>\
+        <body onload='reject()'"
     if request.method == 'POST':
         newItem = Item(
             name=request.form['name'],
             description=request.form['description'],
-            category_id=category_id)
+            category_id=category_id,
+            user_id=category.user_id)
         session.add(newItem)
         session.commit()
         flash('New Item Created')
@@ -252,8 +327,17 @@ def createCategoryItem(category_id):
     '/categories/<int:category_id>/items/<int:item_id>/edit/',
     methods=['GET', 'POST'])
 def editCategoryItem(category_id, item_id):
+    if 'username' not in login_session:
+        return redirect('/login')
+    category = session.query(Category).filter_by(id=category_id).one()
     editedItem = session.query(Item).filter_by(id=item_id).one()
     itemName = editedItem.name
+    if category.user_id != login_session['user_id']:
+        return "\
+        <script>\
+        function reject() {alert('Access denied.');}\
+        </script>\
+        <body onload='reject()'"
     if request.method == 'POST':
         if request.form['name']:
             editedItem.name = request.form['name']
@@ -275,8 +359,17 @@ def editCategoryItem(category_id, item_id):
     '/categories/<int:category_id>/items/<int:item_id>/delete/',
     methods=['GET', 'POST'])
 def deleteCategoryItem(category_id, item_id):
+    if 'username' not in login_session:
+        return redirect('/login')
+    category = session.query(Category).filter_by(id=category_id).one()
     itemToDelete = session.query(Item).filter_by(id=item_id).one()
     itemName = itemToDelete.name
+    if category.user_id != login_session['user_id']:
+        return "\
+        <script>\
+        function reject() {alert('Access denied.');}\
+        </script>\
+        <body onload='reject()'"
     if request.method == 'POST':
         session.delete(itemToDelete)
         session.commit()
